exchanges recreated if vhost is not default /
This should fix issue 173:
https://github.com/puppetlabs/puppetlabs-rabbitmq/issues/173

self.instances would only return the list of exchanges for the default
vhost /, this would cause a problem as any other exchange would be
recreated in every Puppet run.
Also, rabbitmqctl was used instead of rabbitmqadmin because the latest
requires the username and password of the admin user to list the
exchanges, while the first does not.