Remove unused config_mirrored_queues.
Configuring queue mirroring should be done by setting the correct policy
on a queue. Looking at the code this actually hasn't worked in forever,
it's not even doing anything with the variable.

I've deprecated the variable in rabbitmq::server and removed it from
rabbitmq. I doubt anyone was using this anyone as it wasn't doing a
thing.

No tests are affected by this as there aren't any that covered this
attribute.