Replace erlang cookie fact with type and provider
Keeping secrets as facts is a potential security risk since facts
are generally not protected as secrets. This change removes the
rabbitmq_erlang_cookie fact. In order to make this feasible, the
exec that wipes out the rabbitmq db and the file resource managing
the cookie are replaced by a type and provider to reproduce this
functionality.