Fix rabbitmq-env.conf parameters
W/o this patch, env parameters from rabbitmq-env.conf
will always override the ones specified at the init file of
rabbitmq-server.

This is an issue as we may want them to differ
for separated rabbitmq instances running at the same host.

The soution is to remove RABBIT(MQ)_ prefixes for all
parameters in rabbitmq-env.conf as it described at the
man pages.

Closes-bug: #MODULES-1673

Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>