MODULES-1856 - Fix to allow bindings and queues to be created on non-default management port
Without this fix the rabbitmq_binding & rabbitmq_queue providers are unable to work when
rabbitmq is running on a non-default management port. This fix fixes the problem by passing in
the /etc/rabbitmq/rabbitmqadmin.conf to the call to rabbitmqadmin so that the port can be read from the conf file.