Fix future parser errors
Type checking fixes:

Ports should be allowed to be integers, so wrap each validate_re
function call for port-like parameters in `if ! is_integer`.

Scope fixes:

Some parameters were not in scope in the config class, so they were not
available to the rabbitmq.config template. This change adds those
parameters to the config class.