rabbitmqadmin: bring ipv6 support
(1) If node_ip_address is an IPv6, we need to pass specific options to curl
so staging module can work.

When using IPv6, curl needs to be used with specific options, for
example with localhost:
```curl -g -6 [::1]```

(2) Sanitize node_ip_address if IPv6
node_ip_address is used bu rabbitmqadmin.pp and also in config.pp for
rabbitmqenv.config. This parameter should not be given with brackets,
otherwise RabbitMQ won't start.
Though when using curl, we need the add the brackets so this patch makes
sure the brackets are added to node_ip_address if it's an IPv6 so curl
will work correctly.