Extend rabbitmqadmin config template with SSL options.
The rabbitmqadmin CLI tool added better SSL support since version 3.6.2 and allows to set a few more CLI parameter. This puppet module does not incorporate these options when instrumenting the rabbitmqadmin tool. Therefore these options are now added to the config template and set appropriately when ssl is enabled. For earlier versions of RabbitMQ the new options are ignored by the admin tool and should not cause any errors.