[MODULES-4450] don't set ssl depth if undef
As per a recent commit [1], the default value of ssl_depth is undef.
This turned out problematic, as the template checks for 'UNSET', and
validation of this parameter is only done if the ssl_depth was set to
something (remember it defaults to undef). So, using the default value
results in an invalid rabbitmq configuration, which breaks with a syntax
error.

This sets the check in the template to work only if the value of that
variable is set (so it won't write it if the default undef is used).

[1] https://github.com/puppetlabs/puppetlabs-rabbitmq/commit/21ac01e4a1835604e9ffaaedbef024fd2ea9a0f4